1. new fontStyles, add the new fontStyles in homeView. 2. refactor shimmer dish card
