+++ lib/config/themes/font/font_styles.dart
@@ -1,7 +1,19 @@
+import 'package:fast_order/config/themes/index.dart';
 import 'package:flutter/material.dart';
 
-const TextStyle titlOrangeF24 =
-    TextStyle(fontSize: 24, fontWeight: FontWeight.bold, color: Color(0xFFF48C06));
+TextStyle titlOrangeF24 =
+    TextStyle(fontSize: 24, fontWeight: FontWeight.bold, color: colorList[1]);
+
+TextStyle descriptionOrangeF14 =
+    TextStyle(fontSize: 14, fontWeight: FontWeight.bold, color: colorList[1]);
+
+TextStyle subtitleOrangeF16 =
+    TextStyle(fontSize: 16, fontWeight: FontWeight.bold, color: colorList[1]);
+
+TextStyle subtitleOrangeF18 =
+    TextStyle(fontSize: 18, fontWeight: FontWeight.bold, color: colorList[1]);
+
+TextStyle titleOrangeF20 = TextStyle(fontSize: 20, color: colorList[1]);
 
 const TextStyle titleBlackF24 =
     TextStyle(fontSize: 24, fontWeight: FontWeight.bold, color: Colors.black);
@@ -15,12 +27,11 @@ const TextStyle titleWhiteF24 =
 const TextStyle subtitleBlackF16 =
     TextStyle(fontSize: 16, fontWeight: FontWeight.bold, color: Colors.black);
 
+const TextStyle subtitleBlackF18 =
+    TextStyle(fontSize: 18, fontWeight: FontWeight.bold, color: Colors.black);
+
 const TextStyle subtitleBlackF14 = 
     TextStyle(fontSize: 14, fontWeight: FontWeight.bold, color: Colors.black);
 
-const TextStyle descriptionOrangeF14 =
-    TextStyle(fontSize: 14, fontWeight: FontWeight.bold, color: Color(0xFFF48C06));
-
-const TextStyle titleOrangeF20 = TextStyle(fontSize: 20, color: Color(0xFFF48C06));
