+++ lib/presentation/widgets/view/home_view/index.dart
@@ -2,6 +2,5 @@ export './carousel_welcom_box.dart';
 export './welcome_box.dart';
 export './custom_box_date.dart';
 export './available_dishes_carousel.dart';
-export './shimmer_dish_card.dart';
 export 'shimmer_dish_available_text.dart';
 export './dishes_list_view.dart';
